refactoring app component and fixed peer dependency error
