+++ src/setupTests.js
@@ -2,5 +2,3 @@ import '@testing-library/jest-dom'
 import fetchMock from 'jest-fetch-mock'
 
 fetchMock.enableMocks()
-
-global.fetch.moc
