+++ src/styles/movies.scss
@@ -180,6 +180,8 @@
 
 .card {
     position: relative;
+    min-width: 20%;
+    min-height: 100%;
     color: #fff;
     text-align: center;
     background: transparent;
